feat: TypeScript 변화 및 코드 라인 수정 작업
