+++ src/App.styles.tsx
@@ -1,0 +1,45 @@
+import styled from "styled-components";
+
+export const Title = styled.h3``;
+
+export const Container = styled.div`
+  display: flex;
+  background: darkslategray;
+  color: white;
+  flex-direction: column;
+  border: 1px solid gray;
+  border-radius: 20px;
+  margin-bottom: 0.5rem;
+  padding: 1rem;
+`;
+
+export const TitleLogoImage = styled.img`
+  width: 8rem;
+  margin-bottom: 1rem;
+`;
+
+export const Field = styled.div``;
+
+export const Text = styled.span`
+  margin-right: 1rem;
+  display: inline-block;
+  width: 3rem;
+`;
+
+export const Button = styled.button`
+  height: 2rem;
+  background-color: dodgerblue;
+  border-style: none;
+  color: white;
+  border-radius: 10px 10px;
+`;
+
+export const List = styled.div`
+  margin: auto;
+  padding: 1rem;
+  justify-content: center;
+`;
+
+export const A = styled.a`
+  color: crimson;
+`;
