+++ src/components/ReservationList.tsx
@@ -1,0 +1,23 @@
+// import React, { useState } from "react";
+// import { ContainerItem } from "./common/ContainerItem";
+
+// export const ReservationList = () => {
+//   const [reservationList, setReservationList] = useState([]);
+
+//   return (
+//     <ContainerItem
+//       title={"3. 예약 대기 리스트"}
+//       content={
+//         <>
+//           <S.List>
+//             {reservationList &&
+//               reservationList.map((ele, i) => {
+//                 return <p key={i}>{getParseList(ele)}</p>;
+//               })}
+//           </S.List>
+//           <S.Button onClick={summit}>건강 관리실 신청</S.Button>
+//         </>
+//       }
+//     />
+//   );
+// };
