--- src/components/SelectBox.jsx
@@ -1,19 +1,0 @@
-export const SelectBox = (props) => {
-    return (
-            <select
-                name={props.name}
-                onChange={props.onChange}
-                defaultValue={props.options[0]}
-                style={{width:"10.6rem"}}
-            >
-                {props.options.map((option) => (
-                        <option
-                            key={option.value}
-                            value={option.value}
-                        >
-                            {option.name}
-                        </option>
-                ))}
-            </select>
-    );
-};
