+++ src/declare/react-datepicker.d.ts
@@ -1,0 +1,1 @@
+declare module "react-datepicker";
