--- src/utils/Const.js
@@ -1,5 +1,0 @@
-export const Const = {
-    timsUrl: `https://talk.tmaxsoft.com/front/main/main.do`,
-    reservationUrl: `https://talk.tmaxsoft.com/front/health/insertHealth.do?`,
-    historyUrl:`https://talk.tmaxsoft.com/front/health/findHealthResvtList.do?srchMenuNo=TM0041&toggleMenuNo=TM0093`
-}
