+++ src/utils/ConstURL.ts
@@ -1,0 +1,5 @@
+export const constURL = {
+  timsUrl: `https://talk.tmaxsoft.com/front/main/main.do`,
+  reservationUrl: `https://talk.tmaxsoft.com/front/health/insertHealth.do?`,
+  historyUrl: `https://talk.tmaxsoft.com/front/health/findHealthResvtList.do?srchMenuNo=TM0041&toggleMenuNo=TM0093`,
+};
