--- src/utils/FieldNames.js
@@ -1,8 +1,0 @@
-export const fieldNames = {
-    helMngerCd:"관리사",
-    useDate: "날짜",
-    useStTime: "시작 시간",
-    useEdTime: "종료 시간",
-    userDiv: "이용자",
-    reqEmpNo: "사번"
-}
