+++ src/utils/HealthManagerInfo.ts
@@ -1,0 +1,12 @@
+export const healthManagerInfo = [
+  { value: null, name: "선택" },
+  { value: "HEL20231128001", name: "1. 유 정(여/미금)" },
+  { value: "HEL20140101130", name: "2. 이승우(남/미금)" },
+  { value: "HEL20210330001", name: "3. 정영민(남/미금)" },
+  { value: "HEL20231218001", name: "4. 신승란(여/미금)" },
+  { value: "HEL20211216001", name: "5. 이정숙(여/오리)" },
+  { value: "HEL20231220001", name: "6. 김성해(남/오리)" },
+  { value: "HEL20230823001", name: "7. 이정경(여/오리M)" },
+  { value: "HEL20231128004", name: "8. 김도영(여/오리M)" },
+];
+// 예약 가능한 창에서(ex. 다음 달 리스트) findHealthList 소스 확인
