--- src/utils/ParseDate.js
@@ -1,7 +1,0 @@
-export const parseDate = (date) => {
-    return date.toLocaleDateString('ko-KR', {
-        year: 'numeric',
-        month: '2-digit',
-        day: '2-digit'
-    }).replace(/\. /g, '').substring(0, 8);
-}
