+++ src/utils/ParseDate.ts
@@ -1,0 +1,10 @@
+export const parseDate = (date:any) => {
+  return date
+    .toLocaleDateString("ko-KR", {
+      year: "numeric",
+      month: "2-digit",
+      day: "2-digit",
+    })
+    .replace(/\. /g, "")
+    .substring(0, 8);
+};
