--- src/utils/TimeInfo.js
@@ -1,12 +1,0 @@
-export const timeInfo = [
-    { value: null, name: "선택" },
-    { value: "12", name: "12:00 ~ 12:50" },
-    { value: "13", name: "13:00 ~ 13:50" },
-    { value: "14", name: "14:00 ~ 14:50" },
-    { value: "15", name: "15:00 ~ 15:50" },
-    { value: "16", name: "16:00 ~ 16:50" },
-    { value: "17", name: "17:00 ~ 17:50" },
-    { value: "19", name: "19:00 ~ 19:50" },
-    { value: "20", name: "20:00 ~ 20:50" },
-    { value: "21", name: "21:00 ~ 21:50" }
-];
